LU-13224 utils: expose lctl functionality via lustreapi
Add function llapi_param_access() to lustreapi.
This function exposes lctl like functionality for
get_param, set_param, and list_param in for
consumers such as LMT.
The data returned can be written to a file or kept
in memory by the calling application for further
processing. Code was moved from lustre_cfg.c to
liblustreapi.c and lctl is modified to now use lustreapi
for this functionality.

Signed-off-by: Gian-Carlo DeFazio <defazio1@llnl.gov>
Change-Id: I2726b27b0042d58c97284f8348970deb6efc43d1